Refactorment
