+++ packages/npm/@silverback/ui/src/components/1-molecules/__mocks__/mockNavItems.mocks.ts
@@ -1,17 +1,18 @@
 
+import { buildLink } from '@amazeelabs/react-framework-bridge/storybook';
+
 import { NavItem } from '../../../types';
-import { buildLink } from '../../../utils';
 
 export const mockNavItems = (length: number, hasChildren: boolean): Array<NavItem> =>
   Array.from({ length }).map((_, index) => ({
     id: `${index}`,
     name: `Link #${index + 1}`,
-    Link: buildLink(`/go-to-link-${index + 1}`),
+    Link: buildLink({href:`/go-to-link-${index + 1}`}),
     children: hasChildren == true ?
       Array.from({ length }).map((_, index) => ({
         id: `${index}`,
-        name: `SubNavigation Link`,
-        Link: buildLink(`/go-to-link-sub-link`),
+        name: `SubNavigation Link ${index + 1}`,
+        Link: buildLink({href:`/go-to-link-sub-link-${index}`}),
         children: [],
       }))
       : [],
