+++ packages/npm/@silverback/ui/src/components/2-organisms/__stories__/Header.stories.tsx
@@ -1,9 +1,11 @@
-import { Meta } from '@storybook/react';
+import { Meta, StoryObj } from '@storybook/react';
 import { userEvent, within } from '@storybook/testing-library';
-import React from 'react';
+import React, { JSXElementConstructor } from 'react';
 
 import { HeaderMocks } from '../__mocks__/Header.mocks';
 import { Header } from '../Header';
+
+type playStoryType<T extends JSXElementConstructor<any>> = StoryObj<React.ComponentProps<T> &{wouldNavigate: () => void}>;
 
 export default {
   title: 'Components/Organisms/Header',
@@ -13,33 +15,6 @@ export default {
   },
 } as Meta;
 
-export const Default = {
-  render: (args: React.ComponentProps<typeof Header>) => <Header {...HeaderMocks} {...args} />,
-  args: {},
-};
-
-export const ExpandedMenu = {
-  name: 'ExpandedMenu',
-  render: (args: React.ComponentProps<typeof Header>) => <Header {...HeaderMocks} {...args} />,
-  args: {},
-  play: async ({ canvasElement, args }) => {
-    const canvas = within(canvasElement);
-    const mainNavigation = within(
-      await canvas.findByRole('navigation', {
-        name: 'Main navigation',
-      }),
-    );
-    const menuButton = await mainNavigation.findByRole('button', {
-      name: /Link #1/,
-    });
-    await userEvent.click(menuButton);
-
-    await userEvent.click(
-      await within(menuButton.parentElement!).findByRole('link', {
-        name: /^go-to-link-sub-link$/,
-      }),
-    );
-
-    await expect(args.wouldNavigate).toHaveBeenCalled();
-  },
+export const Default: playStoryType<typeof Header> = {
+  args: {navItems: HeaderMocks.navItems, LogoLink: HeaderMocks.LogoLink},
 };
